getComandaById corrigido.

+++ Backend/sb_api/src/main/java/com/snackinback/sb_api/model/Comanda.java
@@ -1,5 +1,6 @@
 package com.snackinback.sb_api.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -41,12 +42,12 @@ public class Comanda {
     @Column(unique = true)
     private String codigoDoPedido;
     private ComandaStatusEnum status;
-    private Double subtotal;
+    private BigDecimal subtotal;
     private LocalDateTime pedidoCriadoEm;
     private LocalDateTime update;
     private MetodoDePagamentoEnum metodoDePagamento;
     
-    public Comanda(List<Item> item, Cliente cliente, String codigoDoPedido, ComandaStatusEnum status, Double subtotal,
+    public Comanda(List<Item> item, Cliente cliente, String codigoDoPedido, ComandaStatusEnum status, BigDecimal subtotal,
             LocalDateTime pedidoCriadoEm, LocalDateTime update, MetodoDePagamentoEnum metodoDePagamento) {
         this.item = item;
         this.cliente = cliente;
@@ -57,5 +58,4 @@ public class Comanda {
         this.update = update;
         this.metodoDePagamento = metodoDePagamento;
     }
-
 }
